服务器开发完善

+++ src/main/java/liar/xiaoyu/www/temp/dao/impl/OrderDaoImpl.java
@@ -15,11 +15,12 @@ public class OrderDaoImpl implements OrderDao {
         PreparedStatement preparedStatement = null;
         ResultSet resultSet = null;
         try {
-            preparedStatement = connection.prepareStatement("SELECT * FROM userorder;");
+            preparedStatement = connection.prepareStatement("SELECT * FROM userorder ORDER BY id DESC;");
             resultSet = preparedStatement.executeQuery();
             while (resultSet.next()){
                 Order order = new Order();
                 order.setId(resultSet.getInt("id"));
+                order.setUuid(resultSet.getString("uuid"));
                 order.setRidgepole(resultSet.getInt("ridgepole"));
                 order.setDorm(resultSet.getInt("dorm"));
                 order.setContact(resultSet.getString("contact"));
@@ -44,6 +45,66 @@ public class OrderDaoImpl implements OrderDao {
             if(preparedStatement!=null){
                 try {
                     preparedStatement.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
+            if(connection!=null){
+                try {
+                    connection.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
+        }
+        return datas;
+    }
+
+    @Override
+    public List<Order> selectAllTemporderByUUID(String uuid) {
+        List<Order> datas = new ArrayList<>();
+        Connection connection = DbcpUtil.getConnection();
+        PreparedStatement preparedStatement = null;
+        ResultSet resultSet = null;
+        try {
+            preparedStatement = connection.prepareStatement("SELECT * FROM userorder WHERE uuid = ? ORDER BY id DESC;");
+            preparedStatement.setString(1,uuid);
+            resultSet = preparedStatement.executeQuery();
+            while (resultSet.next()){
+                Order order = new Order();
+                order.setId(resultSet.getInt("id"));
+                order.setUuid(resultSet.getString("uuid"));
+                order.setRidgepole(resultSet.getInt("ridgepole"));
+                order.setDorm(resultSet.getInt("dorm"));
+                order.setContact(resultSet.getString("contact"));
+                order.setInfo(resultSet.getString("info"));
+                order.setStaff(resultSet.getString("staff"));
+                order.setPhone(resultSet.getString("phone"));
+                order.setFlag(resultSet.getInt("flag"));
+                order.setDatetime(resultSet.getString("datetime"));
+                datas.add(order);
+            }
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }finally {
+            if(resultSet!=null){
+                try {
+                    resultSet.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+
+            }
+            if(preparedStatement!=null){
+                try {
+                    preparedStatement.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
+            if(connection!=null){
+                try {
+                    connection.close();
                 } catch (SQLException e) {
                     e.printStackTrace();
                 }
@@ -59,12 +120,13 @@ public class OrderDaoImpl implements OrderDao {
         PreparedStatement preparedStatement = null;
         ResultSet resultSet = null;
         try {
-            preparedStatement = connection.prepareStatement("SELECT * FROM userorder WHERE ridgepole = ?;");
+            preparedStatement = connection.prepareStatement("SELECT * FROM userorder WHERE ridgepole = ? ORDER BY id DESC;");
             preparedStatement.setInt(1,ridgepole);
             resultSet = preparedStatement.executeQuery();
             while (resultSet.next()){
                 Order order = new Order();
                 order.setId(resultSet.getInt("id"));
+                order.setUuid(resultSet.getString("uuid"));
                 order.setRidgepole(resultSet.getInt("ridgepole"));
                 order.setDorm(resultSet.getInt("dorm"));
                 order.setContact(resultSet.getString("contact"));
@@ -89,6 +151,13 @@ public class OrderDaoImpl implements OrderDao {
             if(preparedStatement!=null){
                 try {
                     preparedStatement.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
+            if(connection!=null){
+                try {
+                    connection.close();
                 } catch (SQLException e) {
                     e.printStackTrace();
                 }
@@ -110,6 +179,7 @@ public class OrderDaoImpl implements OrderDao {
             while (resultSet.next()){
                 order = new Order();
                 order.setId(resultSet.getInt("id"));
+                order.setUuid(resultSet.getString("uuid"));
                 order.setRidgepole(resultSet.getInt("ridgepole"));
                 order.setDorm(resultSet.getInt("dorm"));
                 order.setContact(resultSet.getString("contact"));
@@ -137,6 +207,13 @@ public class OrderDaoImpl implements OrderDao {
                     e.printStackTrace();
                 }
             }
+            if(connection!=null){
+                try {
+                    connection.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
         }
         return order;
     }
@@ -147,15 +224,31 @@ public class OrderDaoImpl implements OrderDao {
         PreparedStatement preparedStatement = null;
         Integer row = 0;
         try {
-            preparedStatement = connection.prepareStatement("DELETE FROM userorder WHERE id = ?;");
+            preparedStatement = connection.prepareStatement("SELECT staff FROM userorder WHERE id = ?;");
             preparedStatement.setInt(1,id);
-            row = preparedStatement.executeUpdate();
+            ResultSet resultSet = preparedStatement.executeQuery();
+            String staff = null;
+            while (resultSet.next()){
+                staff = resultSet.getString("staff");
+            }
+            if(staff == null){
+                preparedStatement = connection.prepareStatement("DELETE FROM userorder WHERE id = ?;");
+                preparedStatement.setInt(1,id);
+                row = preparedStatement.executeUpdate();
+            }
         } catch (SQLException e) {
             e.printStackTrace();
         }finally {
             if(preparedStatement!=null){
                 try {
                     preparedStatement.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
+            if(connection!=null){
+                try {
+                    connection.close();
                 } catch (SQLException e) {
                     e.printStackTrace();
                 }
@@ -170,11 +263,12 @@ public class OrderDaoImpl implements OrderDao {
         PreparedStatement preparedStatement = null;
         Integer id = 0;
         try {
-            preparedStatement = connection.prepareStatement("INSERT INTO userorder(ridgepole,dorm,contact,info) VALUES(?,?,?,?);",Statement.RETURN_GENERATED_KEYS);
-            preparedStatement.setInt(1, order.getRidgepole());
-            preparedStatement.setInt(2, order.getDorm());
-            preparedStatement.setString(3,order.getContact());
-            preparedStatement.setString(4, order.getInfo());
+            preparedStatement = connection.prepareStatement("INSERT INTO userorder(uuid,ridgepole,dorm,contact,info) VALUES(?,?,?,?,?);",Statement.RETURN_GENERATED_KEYS);
+            preparedStatement.setString(1,order.getUuid());
+            preparedStatement.setInt(2, order.getRidgepole());
+            preparedStatement.setInt(3, order.getDorm());
+            preparedStatement.setString(4,order.getContact());
+            preparedStatement.setString(5, order.getInfo());
             Integer row = preparedStatement.executeUpdate();
             if(row > 0){
                 ResultSet generatedKeys = preparedStatement.getGeneratedKeys();
@@ -191,6 +285,13 @@ public class OrderDaoImpl implements OrderDao {
                     e.printStackTrace();
                 }
             }
+            if(connection!=null){
+                try {
+                    connection.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
         }
         return id;
     }
@@ -200,19 +301,37 @@ public class OrderDaoImpl implements OrderDao {
         Connection connection = DbcpUtil.getConnection();
         PreparedStatement preparedStatement = null;
         Integer row = 0;
+        String staff = null;
         try {
-            preparedStatement = connection.prepareStatement("UPDATE userorder SET staff=?,phone=?,flag=? WHERE id = ?;");
-            preparedStatement.setString(1, order.getStaff());
-            preparedStatement.setString(2, order.getPhone());
-            preparedStatement.setInt(3, order.getFlag());
-            preparedStatement.setInt(4, order.getId());
-            row = preparedStatement.executeUpdate();
+            preparedStatement = connection.prepareStatement("SELECT staff FROM userorder WHERE id = ?;");
+            preparedStatement.setInt(1,order.getId());
+            ResultSet resultSet = preparedStatement.executeQuery();
+            while (resultSet.next()){
+                staff = resultSet.getString("staff");
+            }
+            if(staff==null){
+                preparedStatement = connection.prepareStatement("UPDATE userorder SET staff=?,phone=?,flag=? WHERE id = ?;");
+                preparedStatement.setString(1, order.getStaff());
+                preparedStatement.setString(2, order.getPhone());
+                preparedStatement.setInt(3, order.getFlag());
+                preparedStatement.setInt(4, order.getId());
+                row = preparedStatement.executeUpdate();
+            }else{
+                row = 0;
+            }
         } catch (SQLException e) {
             e.printStackTrace();
         }finally {
             if(preparedStatement!=null){
                 try {
                     preparedStatement.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
+            if(connection!=null){
+                try {
+                    connection.close();
                 } catch (SQLException e) {
                     e.printStackTrace();
                 }
